Split GitHub issue filter tests
