+++ src/github/parse.rs
@@ -972,44 +972,3 @@ pub fn build_issue_search_args(
     }
     args
 }
-
-#[cfg(test)]
-mod tests {
-    use super::build_issue_search_args;
-    use crate::domain::IssueFilter;
-
-    fn issue_query_fields(query_arg: &str) -> Vec<&str> {
-        let Some(issue_fields) = query_arg.split("... on Issue {").nth(1) else {
-            return Vec::new();
-        };
-        let Some(before_comments) = issue_fields.split("comments").next() else {
-            return Vec::new();
-        };
-        before_comments
-            .split_whitespace()
-            .filter(|token| token.chars().all(char::is_alphanumeric))
-            .collect()
-    }
-
-    #[test]
-    fn issue_search_args_omit_body_for_fast_first_paint() {
-        let args = build_issue_search_args("owner", "repo", &IssueFilter::default(), None, 30);
-        let query_arg = args
-            .iter()
-            .find(|arg| arg.starts_with("query="))
-            .unwrap_or_else(|| panic!("missing GraphQL query arg: {args:?}"));
-
-        let fields = issue_query_fields(query_arg);
-        assert!(fields.contains(&"title"));
-        assert!(!fields.contains(&"body"));
-
-        let paged_args =
-            build_issue_search_args("owner", "repo", &IssueFilter::default(), Some("cursor"), 30);
-        let paged_query_arg = paged_args
-            .iter()
-            .find(|arg| arg.starts_with("query="))
-            .unwrap_or_else(|| panic!("missing paged GraphQL query arg: {paged_args:?}"));
-        let paged_fields = issue_query_fields(paged_query_arg);
-        assert!(!paged_fields.contains(&"body"));
-    }
-}
